requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Data processing
pandas>=2.0.3
numpy>=1.24.3

# AI/ML
openai>=1.0.0
anthropic>=0.7.0

# File processing
python-magic>=0.4.27
python-docx>=1.0.0
openpyxl>=3.1.2

# Configuration
pyyaml>=6.0.1

# Testing
pytest>=7.4.0

# Cloud and file processing
boto3>=1.28.0
azure-storage-blob>=12.19.0
pdfplumber>=0.10.3
pillow>=10.0.1
pytesseract>=0.3.10
sqlalchemy-utils>=0.41.1
